+++ lib/src/models/models.dart
@@ -4,3 +4,4 @@ export './regions.dart';
 export './states.dart';
 export './brazilian_state.dart';
 export './validation_error.dart';
+export './pix_key_type.dart';